requirements.txt
langchain>=0.2.0
langchain-anthropic>=0.1.0
langgraph>=0.2.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
